Move DBusMenu code to `resources`
Code will be used by other widgets so should be moved out of
the `StatusNotifier` codebase.
